# backend/docs_design/openapi.yaml
...
openapi: 3.1.0
info:
  title: BoThesis Enterprise Agent API
  version: 1.0.0
  description: Contract-first target for workspace knowledge, IAM, integrations, and governed ingestion.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/accounts:
    post:
      tags: [auth]
      security: []
      operationId: createAccount
      description: Create Account and issue its first Session. A guest bearer may be supplied to upgrade an existing guest Session.
      x-optional-bearer: guest-upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AccountCreate'}
      responses:
        '201': {$ref: '#/components/responses/AuthSession'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/auth/sessions:
    post:
      tags: [auth]
      security: []
      operationId: createSession
      description: Create one Session using a typed authentication method. Password, guest, and Google methods are public and may optionally upgrade a guest bearer.
      x-optional-bearer: guest-upgrade
      x-method-security:
        password: public
        guest: public
        google: public
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSessionRequest'}
      responses:
        '200': {$ref: '#/components/responses/AuthSession'}
        '401': {$ref: '#/components/responses/InvalidCredentials'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/AccountDisabled'}
        '429': {$ref: '#/components/responses/TooManyAttempts'}
  /api/v1/auth/session:
    get:
      tags: [auth]
      security:
        - bearerAuth: []
      operationId: getCurrentSession
      responses:
        '200':
          description: Current authenticated Session/context
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    patch:
      tags: [auth]
      security:
        - bearerAuth: []
      operationId: updateCurrentSession
      description: Change the active workspace on the current bearer Session.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CurrentSessionUpdate'}
      responses:
        '200': {$ref: '#/components/responses/AuthSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    delete:
      tags: [auth]
      security:
        - bearerAuth: []
      operationId: deleteCurrentSession
      description: Invalidate the bearer Session resolved from the Authorization header.
      responses:
        '204': {description: Current Session invalidated}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/knowledge/documents/{document_id}:
    get:
      tags: [knowledge]
      operationId: getKnowledgeDocument
      parameters:
        - {$ref: '#/components/parameters/DocumentId'}
        - name: chunk
          in: query
          schema: {type: string, maxLength: 512}
      responses:
        '200':
          description: Permission-checked viewer
          content:
            application/json:
              schema: {$ref: '#/components/schemas/KnowledgeDocumentViewer'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/collections:
    get:
      tags: [collections]
      operationId: listCollections
      x-required-permissions: [knowledge.collections.read]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Search'}
      responses:
        '200':
          description: Permission-filtered Collections
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CollectionPage'}
    post:
      tags: [collections]
      operationId: createCollection
      x-required-permissions: [knowledge.collections.create]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CollectionCreate'}
      responses:
        '201':
          description: Created Collection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Collection'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
  /api/v1/collections/{collection_id}:
    parameters:
      - {$ref: '#/components/parameters/CollectionId'}
    get:
      tags: [collections]
      operationId: getCollection
      responses:
        '200':
          description: Collection detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Collection'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [collections]
      operationId: updateCollection
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CollectionUpdate'}
      responses:
        '200':
          description: Updated Collection
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Collection'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [collections]
      operationId: deleteCollection
      responses:
        '204': {description: Collection tombstoned}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/collections/{collection_id}/documents:
    post:
      tags: [documents]
      operationId: createDocument
      x-required-permissions: [knowledge.documents.create]
      x-required-collection-role: editor
      description: Create one Document. JSON reserves server-owned content storage; multipart stores content directly. Both transports share the same Document lifecycle.
      parameters:
        - {$ref: '#/components/parameters/CollectionId'}
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              {$ref: '#/components/schemas/DocumentMultipartCreate'}
          application/json:
            schema: {$ref: '#/components/schemas/DocumentCreate'}
      responses:
        '200':
          description: Existing Document returned for an idempotent replay
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentCreateResult'}
        '201':
          description: Created Document and optional content instructions
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentCreateResult'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '413': {$ref: '#/components/responses/UploadTooLarge'}
        '415': {$ref: '#/components/responses/UnsupportedMediaType'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/TooManyAttempts'}
  /api/v1/documents:
    get:
      tags: [documents]
      operationId: listDocuments
      x-required-permissions: [knowledge.documents.read]
      x-required-collection-role: viewer
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
        - {$ref: '#/components/parameters/Search'}
        - name: collection_id
          in: query
          schema: {type: string, format: uuid}
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/DocumentStatus'}
      responses:
        '200':
          description: Permission-filtered Documents
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentPage'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/documents/search:
    post:
      tags: [documents]
      operationId: searchDocuments
      x-required-permissions: [knowledge.documents.search]
      x-required-collection-role: viewer
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DocumentSearchRequest'}
      responses:
        '200':
          description: Ranked grounded search results
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentSearchResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/documents/{document_id}:
    parameters:
      - {$ref: '#/components/parameters/DocumentId'}
    get:
      tags: [documents]
      operationId: getDocument
      x-required-permissions: [knowledge.documents.read]
      x-required-collection-role: viewer
      responses:
        '200':
          description: Document detail
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Document'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [documents]
      operationId: deleteDocument
      x-required-permissions: [knowledge.documents.delete]
      x-required-collection-role: editor
      responses:
        '204': {description: Document deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/documents/{document_id}/content:
    put:
      tags: [documents]
      operationId: finalizeDocumentContent
      x-required-permissions: [knowledge.documents.write]
      x-required-collection-role: editor
      description: Validate server-owned object storage content for a reserved Document and make it available.
      parameters:
        - {$ref: '#/components/parameters/DocumentId'}
      responses:
        '200':
          description: Document content is available
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DocumentContentResult'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Search: {name: search, in: query, schema: {type: string, maxLength: 255}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 128}}
...
    CollectionId: {name: collection_id, in: path, required: true, schema: {type: string, format: uuid}}
    DocumentId: {name: document_id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    AuthSession:
      description: Issued access session
      content: {application/json: {schema: {$ref: '#/components/schemas/AuthSession'}}}
    Unauthorized:
      description: Authentication required or invalid
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    InvalidCredentials:
      description: Authentication failed without revealing whether the account or credential was invalid
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Forbidden:
      description: Insufficient permission
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    NotFound:
      description: Resource not found or not visible
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Conflict:
      description: Resource or lifecycle conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    AccountDisabled:
      description: Account is disabled and cannot create a Session
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    ValidationError:
      description: Request validation failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    UploadTooLarge:
      description: Upload exceeds configured limit
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    UnsupportedMediaType:
      description: Unsupported upload content type
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    TooManyAttempts:
      description: Authentication rate limit exceeded
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id, details]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        details: {type: object, additionalProperties: true}
    PageFields:
      type: object
      required: [page, page_size, total]
      properties:
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 100}
        total: {type: integer, minimum: 0}
    AccountCreate:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 8, maxLength: 128}
        username: {type: [string, 'null'], minLength: 3, maxLength: 64}
        display_name: {type: [string, 'null'], maxLength: 255}
    CreateSessionRequest:
      oneOf:
        - {$ref: '#/components/schemas/PasswordSessionCreate'}
        - {$ref: '#/components/schemas/GuestSessionCreate'}
        - {$ref: '#/components/schemas/GoogleSessionCreate'}
      discriminator:
        propertyName: method
        mapping:
          password: '#/components/schemas/PasswordSessionCreate'
          guest: '#/components/schemas/GuestSessionCreate'
          google: '#/components/schemas/GoogleSessionCreate'
    PasswordSessionCreate:
      type: object
      additionalProperties: false
      required: [method, email, password]
      properties:
        method: {type: string, const: password}
        email: {type: string, format: email}
        password: {type: string, format: password, minLength: 8, maxLength: 128, writeOnly: true}
    GuestSessionCreate:
      type: object
      additionalProperties: false
      required: [method]
      properties:
        method: {type: string, const: guest}
    GoogleSessionCreate:
      type: object
      additionalProperties: false
      required: [method, credential]
      properties:
        method: {type: string, const: google}
        credential: {type: string, minLength: 1, maxLength: 12000, writeOnly: true}
    CurrentSessionUpdate:
      type: object
      minProperties: 1
      required: [active_workspace_id]
      properties:
        active_workspace_id: {type: string, format: uuid}
    WorkspaceMembership:
      type: object
      required: [id, code, name, role_codes, permissions]
      properties:
        id: {type: string, format: uuid}
        code: {type: string}
        name: {type: string}
        role_codes: {type: array, items: {type: string}}
        permissions: {type: array, items: {type: string}}
    AuthSession:
      type: object
      required: [access_token, token_type, expires_at, session_id, user_id, active_workspace_id, permissions, platform_permissions, session_kind, workspaces]
      properties:
        access_token: {type: string}
        token_type: {type: string, const: bearer}
        expires_at: {type: string, format: date-time}
        session_id: {type: string, format: uuid}
        user_id: {type: [string, 'null'], format: uuid}
        email: {type: [string, 'null'], format: email}
        display_name: {type: [string, 'null']}
        active_workspace_id: {type: string, format: uuid}
        permissions: {type: array, items: {type: string}}
        platform_permissions: {type: array, items: {type: string}}
        session_kind: {type: string, enum: [user, guest]}
        workspaces: {type: array, items: {$ref: '#/components/schemas/WorkspaceMembership'}}
    CurrentSession:
      type: object
      required: [session_id, expires_at, user_id, active_workspace_id, permissions, platform_permissions, session_kind, workspaces]
      properties:
        session_id: {type: string, format: uuid}
        expires_at: {type: string, format: date-time}
        user_id: {type: [string, 'null'], format: uuid}
        email: {type: [string, 'null'], format: email}
        display_name: {type: [string, 'null']}
        active_workspace_id: {type: string, format: uuid}
        permissions: {type: array, items: {type: string}}
        platform_permissions: {type: array, items: {type: string}}
        session_kind: {type: string, enum: [user, guest]}
        workspaces: {type: array, items: {$ref: '#/components/schemas/WorkspaceMembership'}}
...
    CollectionStatus: {type: string, enum: [active, archived]}
    Collection:
      type: object
      required: [id, title, status, document_count, source_count, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        description: {type: [string, 'null']}
        parent_collection_id: {type: [string, 'null'], format: uuid}
        status: {$ref: '#/components/schemas/CollectionStatus'}
        document_count: {type: integer, minimum: 0}
        source_count: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CollectionCreate:
      type: object
      required: [title]
      properties:
        title: {type: string, minLength: 1, maxLength: 255}
        description: {type: [string, 'null'], maxLength: 2000}
        parent_collection_id: {type: [string, 'null'], format: uuid}
        inherit_access: {type: boolean, default: true}
    CollectionUpdate:
      type: object
      minProperties: 1
      properties:
        title: {type: string, minLength: 1, maxLength: 255}
        description: {type: [string, 'null'], maxLength: 2000}
        status: {$ref: '#/components/schemas/CollectionStatus'}
    CollectionPage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/Collection'}}}
...
    DocumentStatus: {type: string, enum: [pending_content, available, failed]}
    Document:
      type: object
      additionalProperties: false
      required: [id, collection_id, name, content_type, size_bytes, purpose, status, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        collection_id: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 240}
        content_type: {type: string}
        size_bytes: {type: integer, minimum: 0}
        purpose: {type: string, enum: [knowledge, conversation_attachment]}
        status: {$ref: '#/components/schemas/DocumentStatus'}
        latest_ingestion_id: {type: [string, 'null'], format: uuid}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    DocumentPage:
      allOf:
        - {$ref: '#/components/schemas/PageFields'}
        - type: object
          required: [items]
          properties: {items: {type: array, items: {$ref: '#/components/schemas/Document'}}}
    DocumentSearchRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 512}
        top_k: {type: integer, minimum: 1, maximum: 20, default: 6}
        collection_ids: {type: array, maxItems: 20, uniqueItems: true, items: {type: string, format: uuid}}
    DocumentSearchResult:
      type: object
      additionalProperties: false
      required: [document_id, collection_id, name, excerpt, score]
      properties:
        document_id: {type: string, format: uuid}
        collection_id: {type: string, format: uuid}
        name: {type: string}
        excerpt: {type: string}
        score: {type: number}
        url: {type: [string, 'null'], format: uri}
        metadata: {type: object, additionalProperties: true}
    DocumentSearchResponse:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/DocumentSearchResult'}}
        total: {type: integer, minimum: 0}
    DocumentCreate:
      type: object
      additionalProperties: false
      required: [name, content_type, size_bytes]
      properties:
        name: {type: string, minLength: 1, maxLength: 240}
        content_type: {type: string, minLength: 1, maxLength: 160}
        size_bytes: {type: integer, minimum: 1}
        purpose: {type: string, enum: [knowledge, conversation_attachment], default: knowledge}
    DocumentMultipartCreate:
      type: object
      additionalProperties: false
      required: [file]
      properties:
        file: {type: string, format: binary}
        purpose: {type: string, enum: [knowledge, conversation_attachment], default: knowledge}
    DocumentContentInstructions:
      type: object
      additionalProperties: false
      required: [url, method, headers, expires_at]
      properties:
        url: {type: string, format: uri}
        method: {type: string, enum: [PUT]}
        headers: {type: object, additionalProperties: {type: string}}
        expires_at: {type: string, format: date-time}
    DocumentCreateResult:
      type: object
      additionalProperties: false
      required: [document, upload, ingestion, created]
      properties:
        document: {$ref: '#/components/schemas/Document'}
        upload: {anyOf: [{$ref: '#/components/schemas/DocumentContentInstructions'}, {type: 'null'}]}
        ingestion: {anyOf: [{$ref: '#/components/schemas/Ingestion'}, {type: 'null'}]}
        created:
          type: boolean
          description: True only when this request created the Document; false on an idempotent replay.
    DocumentContentResult:
      type: object
      additionalProperties: false
      required: [document, ingestion]
      properties:
        document: {$ref: '#/components/schemas/Document'}
        ingestion: {anyOf: [{$ref: '#/components/schemas/Ingestion'}, {type: 'null'}]}
...
    KnowledgeDocumentViewer:
      type: object
      required: [document_id, title, content_type, status, elements]
      properties:
        document_id: {type: string, format: uuid}
        title: {type: string}
        content_type: {type: string}
        status: {$ref: '#/components/schemas/DocumentStatus'}
        document_url: {type: [string, 'null'], format: uri}
        external_url: {type: [string, 'null'], format: uri}
        elements: {type: array, items: {type: object, additionalProperties: true}}
...
    IngestionStatus: {type: string, enum: [pending, running, completed, failed, cancelled, timed_out]}
    Ingestion:
      type: object
      required: [id, status, trigger_type, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        source_id: {type: [string, 'null'], format: uuid}
        document_id: {type: [string, 'null'], format: uuid}
        connection_id: {type: [string, 'null'], format: uuid}
        status: {$ref: '#/components/schemas/IngestionStatus'}
        trigger_type: {type: string, enum: [manual, scheduled, webhook, initial, upload, retry]}
        retry_of_ingestion_id: {type: [string, 'null'], format: uuid}
        progress:
          type: object
          properties:
            phase: {type: string}
            discovered_count: {type: integer, minimum: 0}
            processed_count: {type: integer, minimum: 0}
            indexed_count: {type: integer, minimum: 0}
            deleted_count: {type: integer, minimum: 0}
            failed_count: {type: integer, minimum: 0}
        started_at: {type: [string, 'null'], format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...